refactor chain collectors and internal contexts
