+++ common/changes/@itwin/core-geometry/da4-multichaincollector-refactor_2023-07-14-22-29.json
@@ -1,0 +1,10 @@
+{
+  "changes": [
+    {
+      "packageName": "@itwin/core-geometry",
+      "comment": "refactor chain collectors and internal contexts",
+      "type": "none"
+    }
+  ],
+  "packageName": "@itwin/core-geometry"
+}
