+++ core/geometry/src/curve/CurvePrimitive.ts
@@ -6,36 +6,36 @@
  * @module Curve
  */
 import { assert } from "@itwin/core-bentley";
-import { Clipper } from "../clipping/ClipUtils";
 import { StrokeCountMap } from "../curve/Query/StrokeCountMap";
-import { AxisOrder, Geometry, PlaneAltitudeEvaluator } from "../Geometry";
-import { IStrokeHandler } from "../geometry3d/GeometryHandler";
+import { AxisOrder, Geometry, type PlaneAltitudeEvaluator } from "../Geometry";
 import { Matrix3d } from "../geometry3d/Matrix3d";
-import { Plane3dByOriginAndUnitNormal } from "../geometry3d/Plane3dByOriginAndUnitNormal";
-import { Plane3dByOriginAndVectors } from "../geometry3d/Plane3dByOriginAndVectors";
-import { Point3d, Vector3d } from "../geometry3d/Point3dVector3d";
-import { Range1d, Range3d } from "../geometry3d/Range";
+import { Point3d, type Vector3d } from "../geometry3d/Point3dVector3d";
+import { type Range1d, Range3d } from "../geometry3d/Range";
 import { Ray3d } from "../geometry3d/Ray3d";
 import { Transform } from "../geometry3d/Transform";
-import { VariantCurveExtendParameter } from "./CurveExtendMode";
 import { CurveIntervalRole, CurveLocationDetail, CurveSearchStatus } from "./CurveLocationDetail";
 import { GeometryQuery } from "./GeometryQuery";
 import { AppendPlaneIntersectionStrokeHandler } from "./internalContexts/AppendPlaneIntersectionStrokeHandler";
 import { ClosestPointStrokeHandler } from "./internalContexts/ClosestPointStrokeHandler";
 import { CurveLengthContext } from "./internalContexts/CurveLengthContext";
 import { LineString3d } from "./LineString3d";
-import { StrokeOptions } from "./StrokeOptions";
 
-import type { InterpolationCurve3d } from "../bspline/InterpolationCurve3d";
 import type { AkimaCurve3d } from "../bspline/AkimaCurve3d";
-import type { BSplineCurve3d } from "../bspline/BSplineCurve";
+import type { Arc3d } from "./Arc3d";
 import type { BezierCurve3d } from "../bspline/BezierCurve3d";
+import type { BSplineCurve3d } from "../bspline/BSplineCurve";
+import type { Clipper } from "../clipping/ClipUtils";
 import type { CurveChainWithDistanceIndex } from "./CurveChainWithDistanceIndex";
 import type { DirectSpiral3d } from "./spiral/DirectSpiral3d";
 import type { IntegratedSpiral3d } from "./spiral/IntegratedSpiral3d";
-import type { Arc3d } from "./Arc3d";
+import type { InterpolationCurve3d } from "../bspline/InterpolationCurve3d";
+import type { IStrokeHandler } from "../geometry3d/GeometryHandler";
 import type { LineSegment3d } from "./LineSegment3d";
-import type { OffsetOptions } from "./internalContexts/PolygonOffsetContext";
+import type { OffsetOptions } from "./OffsetOptions";
+import type { Plane3dByOriginAndUnitNormal } from "../geometry3d/Plane3dByOriginAndUnitNormal";
+import type { Plane3dByOriginAndVectors } from "../geometry3d/Plane3dByOriginAndVectors";
+import type { StrokeOptions } from "./StrokeOptions";
+import type { VariantCurveExtendParameter } from "./CurveExtendMode";
 
 /**
  * Describes the concrete type of a [[CurvePrimitive]]. Each type name maps to a specific subclass and can be used
