+++ core/geometry/src/curve/spiral/TransitionSpiral3d.ts
@@ -15,8 +15,8 @@ import { Transform } from "../../geometry3d/Transform";
 import { CurvePrimitive } from "../CurvePrimitive";
 import { CurveOffsetXYHandler } from "../internalContexts/CurveOffsetXYHandler";
 import { PlaneAltitudeRangeContext } from "../internalContexts/PlaneAltitudeRangeContext";
-import { OffsetOptions } from "../internalContexts/PolygonOffsetContext";
 import { LineString3d } from "../LineString3d";
+import { OffsetOptions } from "../OffsetOptions";
 import { TransitionConditionalProperties } from "./TransitionConditionalProperties";
 
 import type { Vector3d } from "../../geometry3d/Point3dVector3d";
